src/Elksoft.Converters/Elksoft.Converters.ImplicitNumericConverters/ImplicitCharToUIntPtrConverter.cs
// <copyright file="ImplicitCharToUIntPtrConverter.cs" company="Georgy Losenkov">
// Copyright (c) Georgy Losenkov. All rights reserved.
// </copyright>

#if NET7_0_OR_GREATER
using System;

namespace Elksoft.Converters.ImplicitNumericConverters
{
    internal sealed class ImplicitCharToUIntPtrConverter : Converter<Char, UIntPtr>
    {
        public override Boolean AcceptsNull => false;

        public override Boolean IsExplicit => false;

        public override UIntPtr Convert(Char value, IFormatProvider? formatProvider)
        {
            return unchecked((UIntPtr)value);
        }
    }
}
#endif
